Add authorization to the endpoints

+++ src/main/java/org/cru/webservices/DeleteResource.java
@@ -6,6 +6,7 @@ import org.cru.model.Person;
 import org.cru.model.SearchResponse;
 import org.cru.qualifiers.Delete;
 import org.cru.qualifiers.Match;
+import org.cru.service.AuthService;
 import org.cru.service.DeleteService;
 
 import javax.inject.Inject;
@@ -15,6 +16,8 @@ import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.Produces;
 import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.net.ConnectException;
@@ -38,13 +41,16 @@ public class DeleteResource
     private MatchingService matchingService;
     @Inject
     private PersonDeserializer personDeserializer;
+    @Inject
+    private AuthService authService;
 
     @SuppressWarnings("unused")  //used by Clients
     @Path("/delete/{id}")
     @DELETE
     @Produces(MediaType.APPLICATION_JSON)
-    public Response deletePersonByGlobalRegistryId(String globalRegistryId)
+    public Response deletePersonByGlobalRegistryId(String globalRegistryId, @Context HttpHeaders httpHeaders)
     {
+        if(!authService.hasAccess(httpHeaders)) return authService.notAuthorized(httpHeaders);
         deleteService.deletePerson(globalRegistryId, matchingService.findMatchInMdmByGlobalRegistryId(globalRegistryId));
         return Response.ok().entity(buildResponseEntity(globalRegistryId)).build();
     }
